requirements.txt
langchain
langgraph
langchain-openai
langchain-community
tavily-python
arxiv
faiss-cpu
pypdf
python-dotenv
matplotlib
seaborn
pandas
tiktoken
langchain-experimental
langchain-huggingface
sentence-transformers
pdf2image
pytesseract
pillow
